Fix a bunch of build errors
