+++ src/Pages/OATEditorPage/OATEditorPage.state.ts
@@ -23,7 +23,7 @@ import {
     getStoredEditorModelsData,
     getStoredEditorNamespaceData,
     getStoredEditorTemplateData
-} from '../../Models/Services/Utils';
+} from '../../Models/Services/OatUtils';
 
 export const defaultOATEditorState: IOATEditorState = {
     selection: null,
@@ -74,7 +74,6 @@ export const OATEditorPageReducer = produce(
                 state.templates = payload.templates;
                 state.namespace = payload.namespace;
                 state.modelsMetadata = payload.modelsMetadata;
-                state.model = null;
                 return;
             case SET_OAT_PROJECT_NAME:
                 state.projectName = payload;
